// temper/shared-addon-libraries-lib-data-encode/src/types.ts
export type LdeValue = string | number

export type LuaTable = Record<string | number, unknown>
// ...
export interface EncodeInstance {
  data: unknown
  encodedStrings: string[]
  currentString: string
  currentStringLength: number
  globalDictionary: LdeValue[] | undefined
  dictionary: LdeValue[]
  reverseDictionary: LuaTable | undefined
  Initialize: (
    this: EncodeInstance,
    data: unknown,
    localDictionary?: LdeValue[] | true,
    globalDictionary?: LdeValue[]
  ) => void
  InitDictionary: (this: EncodeInstance, localDictionary: LdeValue[] | true) => void
  MakeReverseDictionary: (this: EncodeInstance) => void
  AddString: (this: EncodeInstance, str: string) => void
  AddInteger: (this: EncodeInstance, integer: number) => void
  NewLine: (this: EncodeInstance) => void
  EncodeDictionary: (this: EncodeInstance, dictionary: LdeValue[]) => void
  CheckForStringId: (this: EncodeInstance, value: unknown) => number | undefined
  EncodeItem: (this: EncodeInstance, value: unknown) => void
  EncodeArray: (this: EncodeInstance, array: unknown) => void
  EncodeTable: (this: EncodeInstance, table: unknown) => void
  [key: string]: unknown
}

export interface EncodeClass {
  Subclass: (this: EncodeClass) => EncodeClass
  New: (
    this: EncodeClass,
    data: unknown,
    localDictionary?: LdeValue[] | true,
    globalDictionary?: LdeValue[]
  ) => EncodeInstance
  Initialize: (
    this: EncodeInstance,
    data: unknown,
    localDictionary?: LdeValue[] | true,
    globalDictionary?: LdeValue[]
  ) => void
  InitDictionary: (this: EncodeInstance, localDictionary: LdeValue[] | true) => void
  MakeReverseDictionary: (this: EncodeInstance) => void
  AddString: (this: EncodeInstance, str: string) => void
  AddInteger: (this: EncodeInstance, integer: number) => void
  NewLine: (this: EncodeInstance) => void
  EncodeDictionary: (this: EncodeInstance, dictionary: LdeValue[]) => void
  CheckForStringId: (this: EncodeInstance, value: unknown) => number | undefined
  EncodeItem: (this: EncodeInstance, value: unknown) => void
  EncodeArray: (this: EncodeInstance, array: unknown) => void
  EncodeTable: (this: EncodeInstance, table: unknown) => void
  [key: string]: unknown
}